Move replication data handling out of generic_worker
This is to reduce the difference between master and other workers.
